+++ src/services/index.ts
@@ -1,5 +1,5 @@
 export * from './alkemio-client-adapter';
 export * from './application-notification-builder/application.notification.builder';
 export * from './notification/notification.service';
-export * from './notification-receiver-yml/notification.receivers.yml.service';
-export * from './notification-receivers-template-yml/notification.receivers.template.yml.service';
+export * from './template-to-credential-mapper/template.to.credential.mapper';
+export * from './notification-recipients-adapter/notification.recipient.yml.adapter';
