--- src/services/notification-receiver-yml/index.ts
@@ -1,1 +1,0 @@
-export * from './notification.receivers.yml.service';
