--- src/services/notification-receivers-template-yml/index.ts
@@ -1,1 +1,0 @@
-export * from './notification.receivers.template.yml.service';
