+++ src/services/notification-recipients-adapter/index.ts
@@ -1,0 +1,1 @@
+export * from './notification.recipient.yml.adapter';
