+++ src/services/template-to-credential-mapper/index.ts
@@ -1,0 +1,1 @@
+export * from './template.to.credential.mapper';
