+++ src/services/template-to-credential-mapper/utils/utils.ts
